+++ packages/components/badge/src/Badge/Badge.styles.ts
@@ -79,6 +79,9 @@ export const getBadgeStyles = ({
     letterSpacing:
       '0.06rem' /*move to tokens or update wide letter spacing token*/,
     borderRadius: `${tokens.borderRadiusSmall}`,
+    whiteSpace: 'nowrap',
+    overflow: 'hidden',
+    textOverflow: 'ellipsis',
     ...variantToStyles({ variant }),
     ...sizeToStyles({ size }),
   });