+++ packages/components/button/src/ToggleButton/ToggleButton.test.tsx
@@ -6,22 +6,32 @@ import { axe } from '@/scripts/test/axeHelper';
 
 import { ToggleButton } from '.';
 describe('ToggleButton', function () {
+  const mockOnToggle = jest.fn();
+
   it('renders the component', () => {
-    render(<ToggleButton>Toggle</ToggleButton>);
+    render(<ToggleButton onToggle={mockOnToggle}>Toggle</ToggleButton>);
 
     expect(screen.getByRole('button')).toBeTruthy();
   });
 
   it('renders the component with an additional class name', () => {
     const additionalClassName = 'my-extra-class';
-    render(<ToggleButton className={additionalClassName}>Toggle</ToggleButton>);
+    render(
+      <ToggleButton onToggle={mockOnToggle} className={additionalClassName}>
+        Toggle
+      </ToggleButton>,
+    );
 
     const button = screen.getByRole('button');
     expect(button.classList.contains(additionalClassName)).toBeTruthy();
   });
 
   it('renders the component active', () => {
-    render(<ToggleButton isActive>Toggle</ToggleButton>);
+    render(
+      <ToggleButton onToggle={mockOnToggle} isActive>
+        Toggle
+      </ToggleButton>,
+    );
 
     const button = screen.getByRole('button');
     expect(button.getAttribute('aria-pressed')).toBe('true');
@@ -29,15 +39,17 @@ describe('ToggleButton', function () {
   });
 
   it('renders the component with icon', () => {
-    render(<ToggleButton icon={<PreviewIcon />}>Toggle</ToggleButton>);
+    render(
+      <ToggleButton onToggle={mockOnToggle} icon={<PreviewIcon />}>
+        Toggle
+      </ToggleButton>,
+    );
 
     const button = screen.getByRole('button');
     expect(button.getElementsByTagName('svg')).toHaveLength(1);
   });
 
   it('should not dispatch onClick if disabled', () => {
-    const mockOnToggle = jest.fn();
-
     render(
       <ToggleButton onToggle={mockOnToggle} icon={<PreviewIcon />} isDisabled>
         Toggle
@@ -50,7 +62,9 @@ describe('ToggleButton', function () {
   });
 
   it('has no a11y issues', async () => {
-    const { container } = render(<ToggleButton>Toggle</ToggleButton>);
+    const { container } = render(
+      <ToggleButton onToggle={mockOnToggle}>Toggle</ToggleButton>,
+    );
     const results = await axe(container);
 
     expect(results).toHaveNoViolations();