+++ packages/components/notification/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.14](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.13...@contentful/f36-notification@4.0.14) (2022-02-22)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
+## [4.0.13](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.12...@contentful/f36-notification@4.0.13) (2022-02-22)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
+## [4.0.12](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.11...@contentful/f36-notification@4.0.12) (2022-02-17)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
 ## [4.0.11](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.10...@contentful/f36-notification@4.0.11) (2022-02-09)
 
 **Note:** Version bump only for package @contentful/f36-notification