+++ packages/components/typography/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.2.0](https://github.com/contentful/forma-36/compare/@contentful/f36-typography@4.1.2...@contentful/f36-typography@4.2.0) (2022-02-22)
+
+
+### Features
+
+* new property to typography components ([#1891](https://github.com/contentful/forma-36/issues/1891)) ([ed309a6](https://github.com/contentful/forma-36/commit/ed309a69ec840b0c0b38150403d2839a2678ad9e))
+
+
+
+
+
+## [4.1.2](https://github.com/contentful/forma-36/compare/@contentful/f36-typography@4.1.1...@contentful/f36-typography@4.1.2) (2022-02-17)
+
+**Note:** Version bump only for package @contentful/f36-typography
+
+
+
+
+
 ## [4.1.1](https://github.com/contentful/forma-36/compare/@contentful/f36-typography@4.1.0...@contentful/f36-typography@4.1.1) (2022-01-31)
 
 **Note:** Version bump only for package @contentful/f36-typography